% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{beanie}
\alias{beanie}
\title{Age and value of 50 beanie baby toys}
\format{
An object of class \code{data.frame} with 50 rows and 3 columns.
}
\source{
Beanie world magazine
 @format A data frame with 50 rows and 3 columns:
\describe{
  \item{name}{Name of the toy}
  \item{age}{Age of the toy in months}
  \item{value}{Market value of the toy in US dollars}
}
}
\usage{
beanie
}
\description{
Age and value of 50 beanie baby toys
}
\examples{
 summary(beanie)
 head(beanie)
}
\keyword{datasets}
