% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{cement}
\alias{cement}
\title{Breaking strength of cement data}
\format{
A data frame with 36 rows and 3 columns:
\describe{
 \item{strength}{Breaking strength in pounds per square inch}
 \item{gauger}{Three different gauger machines which mixes cement with water}
 \item{breaker}{Three different breakers breaking the cement cubes}
}
}
\usage{
cement
}
\description{
Breaking strength of cement data
}
\examples{
 summary(cement)
}
\keyword{datasets}
