% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clt_illustrates.R
\name{see_the_clt_for_Bernoulli}
\alias{see_the_clt_for_Bernoulli}
\title{Illustration of the CLT for samples from the Bernoulli distribution}
\usage{
see_the_clt_for_Bernoulli(nsize = 10, nrep = 10000, prob = 0.8)
}
\arguments{
\item{nsize}{Sample size, n. Its default value is 10.}

\item{nrep}{Number of replications. How many samples of size \code{nsize} 
should be taken, default value is 10000.}

\item{prob}{True probability of success for the Bernoulli trials}
}
\description{
Illustration of the CLT for samples from the Bernoulli distribution
}
\examples{
see_the_clt_for_Bernoulli()
par(mfrow=c(2, 3))
see_the_clt_for_Bernoulli(nsize=30)
see_the_clt_for_Bernoulli(nsize=50)
see_the_clt_for_Bernoulli(nsize=100)
see_the_clt_for_Bernoulli(nsize=500)
see_the_clt_for_Bernoulli(nsize=1000)
see_the_clt_for_Bernoulli(nsize=5000)
par(mfrow=c(1,1))
}
