% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clt_illustrates.R
\name{see_the_clt_for_uniform}
\alias{see_the_clt_for_uniform}
\title{Illustration of the central limit theorem for sampling from the 
uniform distribution}
\usage{
see_the_clt_for_uniform(nsize = 10, nrep = 10000)
}
\arguments{
\item{nsize}{Sample size, n. Its default value is 10.}

\item{nrep}{Number of replications. How many samples of size \code{nsize} 
should be taken, default value is 10000.}
}
\value{
A vector of means of the replicated samples.  
The function also illustrates using a graph.
}
\description{
Illustration of the central limit theorem for sampling from the 
uniform distribution
}
\examples{
 see_the_clt_for_uniform()
par(mfrow=c(2, 3))
a1 <- see_the_clt_for_uniform(nsize=1)
a2 <- see_the_clt_for_uniform(nsize=2)
a3 <- see_the_clt_for_uniform(nsize=5)
a4 <- see_the_clt_for_uniform(nsize=10)
a5 <- see_the_clt_for_uniform(nsize=20)
a6 <- see_the_clt_for_uniform(nsize=50)
par(mfrow=c(1, 1))

}
